just practicing oop approach in cpp
