+++ resources_examples/example.cpp
