+++ resources_examples/test.cpp
